feat: validacoes
esqueleto de métodos para validar ortogonalidade e ortonormalidade, ainda com bugs
